# cmd/docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "GoFundIt is a collaborative platform enabling families and friends to plan and fund group activities, such as vacations and events. Features end-to-end encryption, secure payment handling, and flexible contribution options.",
        "title": "GoFundIt API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "https://gofundit.com/support",
            "email": "oyeniyibright@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/activity/{campaignID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all activities for a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get Activities",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activities fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Activity"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Campaign not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new activity for a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Create Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Activity Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Activity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid inputs",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/activity/{campaignID}/{activityID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific activity by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Activity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Activity ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    },
                    "404": {
                        "description": "Activity not found",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes an activity from a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Delete Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Activity ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Update Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update Activity Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateActivityRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Activity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid inputs",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.BadRequestResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.ValidationError"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
        "/activity/{campaignID}/{activityID}/approve": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Approves an activity in a campaign",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Approve Activity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Activity approved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Activity"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Activity ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
...
        "/activity/{campaignID}/{activityID}/participants": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all participants for an activity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "activity"
                ],
                "summary": "Get Participants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Campaign ID",
                        "name": "campaignID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Activity ID",
                        "name": "activityID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Participants fetched successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Contributor"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Activity ID",
                        "schema": {
                            "$ref": "#/definitions/handlers.BadRequestResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.UnauthorizedResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.BadRequestResponse": {
            "description": "Bad request error response structure",
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Validation errors (optional)\n@example [{\"field\":\"field\",\"error\":\"must be a valid data\"}]"
                },
                "message": {
                    "description": "Error message\n@example Invalid input provided",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"Bad Request\"\n@example Bad Request",
                    "type": "string",
                    "enum": [
                        "Bad Request"
                    ],
                    "example": "Bad Request"
                }
            }
        },
        "handlers.SuccessResponse": {
            "description": "Successful API response structure",
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)\n@example null"
                },
                "message": {
                    "description": "Response message describing the success\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Status will always be \"OK\"\n@example OK",
                    "type": "string",
                    "enum": [
                        "OK"
                    ],
                    "example": "OK"
                }
            }
        },
        "handlers.UnauthorizedResponse": {
            "description": "Unauthorized error response structure",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Will always be \"Unauthorized\"\n@example Unauthorized",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                },
                "status": {
                    "description": "Status will always be false for unauthorized\n@example false",
                    "type": "string",
                    "enum": [
                        "Unauthorized"
                    ],
                    "example": "Unauthorized"
                }
            }
        },
...
        "handlers.response": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data (optional)"
                },
                "errors": {
                    "description": "Error details (optional)"
                },
                "message": {
                    "description": "Response message\n@example Operation completed successfully",
                    "type": "string"
                },
                "status": {
                    "description": "Response status\n@example OK",
                    "type": "string"
                }
            }
        },
        "models.Activity": {
            "type": "object",
            "required": [
                "campaignId",
                "cost",
                "title"
            ],
            "properties": {
                "campaignId": {
                    "type": "string"
                },
                "contributors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Contributor"
                    }
                },
                "cost": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isApproved": {
                    "type": "boolean"
                },
                "isMandatory": {
                    "type": "boolean"
                },
                "subtitle": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for accessing the API endpoints",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "JWT token for authenticated requests. Use the format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CampaignKeyAuth": {
            "description": "Campaign Key required for campaign-specific operations",
            "type": "apiKey",
            "name": "Campaign-Key",
            "in": "header"
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}